requirements.txt
python-telegram-bot==21.3
openai>=1.54.0
numpy==1.26.4
python-dotenv==1.0.1
tiktoken==0.7.0
